LayoutTests/imported/w3c/web-platform-tests/html/canvas/historical.any.worker-expected.txt


PASS OffscreenCanvasRenderingContext2D.commit method is removed
